+++ Tests/BypassFinalHook.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the ekino/data-protection-bundle project.
+ *
+ * (c) Ekino
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Ekino\DataProtectionBundle\Tests;
+
+use DG\BypassFinals;
+use PHPUnit\Runner\BeforeTestHook;
+
+final class BypassFinalHook implements BeforeTestHook
+{
+    public function executeBeforeTest(string $test): void
+    {
+        BypassFinals::enable();
+    }
+}